feat: add cart button to resource overview
